Remove all instances of long numeric literals.

+++ IPython/utils/traitlets.py
@@ -896,7 +896,7 @@ else:
     class Long(TraitType):
         """A long integer trait."""
 
-        default_value = 0L
+        default_value = long(0)
         info_text = 'a long'
 
         def validate(self, obj, value):
